+++ sekoia-io-xdr/constants.py
@@ -5,5 +5,6 @@ BASE_URL = "https://app.sekoia.io"
 ALERTS_BASE_URL = f"{BASE_URL}/api/v1/sic/alerts"
 
 ASSETS_BASE_URL = f"{BASE_URL}/api/v1/asset-management/assets"
+ASSETS_V2_BASE_URL = f"{BASE_URL}/api/v2/asset-management/assets"
 
 CASES_BASE_URL = f"{BASE_URL}/api/v1/sic/cases"